[SC-2461] Extract secret redaction into internal/redact
The hook-event pipeline must scrub tool inputs on the same terms the
diagnosis path already applies to tracker comments, but internal/agent
imports internal/daemon, so the daemon cannot import the redaction that
lives in diagnose.go without a cycle. Move the token-shape regexes,
env-derived secret collection, and home-dir de-identification into a
neutral stdlib-only internal/redact package both sides depend on, and
delegate diagnose.go to it. This is the reuse the fix needs, not a copy.
